IT WORKS WITH DOUBLESgit add -Agit add -A

+++ header/tokenization.h
@@ -21,7 +21,7 @@ typedef enum TokenType {
 typedef struct Token {
     TokenType type;
     Line name;
-    long number; 
+    double number; 
 } Token;
 define_array(Token);
